requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
openai>=1.30.0
markitdown[all]>=0.0.1a2
pandas>=2.2.0
openpyxl>=3.1.2
python-docx>=1.1.0
Jinja2>=3.1.4
aiofiles>=23.2.1
httpx>=0.27.0
pillow>=10.3.0
